Move logout pages to seperate folder
