+++ src/containers/Logout/LogoutFederated.js
@@ -1,20 +1,19 @@
 /**
- * Copyright (c) 2016-present, NDLA.
+ * Copyright (C) 2017 -present, NDLA
  *
  * This source code is licensed under the GPLv3 license found in the
  * LICENSE file in the root directory of this source tree.
- *
  */
 
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import * as actions from './sessionActions';
+import * as actions from '../App/sessionActions';
 
-export class LogoutSession extends React.Component {
+export class LogoutFederated extends React.Component {
   componentWillMount() {
     const { logout } = this.props;
-    logout({ federated: false });
+    logout({ federated: true });
   }
 
   render() {
@@ -22,12 +21,13 @@ export class LogoutSession extends React.Component {
   }
 }
 
-LogoutSession.propTypes = {
+LogoutFederated.propTypes = {
   logout: PropTypes.func.isRequired,
+  authenticated: PropTypes.func.isRequired,
 };
 
 const mapDispatchToProps = {
   logout: actions.logout,
 };
 
-export default connect(undefined, mapDispatchToProps)(LogoutSession);
+export default connect(undefined, mapDispatchToProps)(LogoutFederated);
