+++ src/containers/Logout/LogoutSession.js
@@ -1,19 +1,20 @@
 /**
- * Copyright (C) 2017 -present, NDLA
+ * Copyright (c) 2016-present, NDLA.
  *
  * This source code is licensed under the GPLv3 license found in the
  * LICENSE file in the root directory of this source tree.
+ *
  */
 
 import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import * as actions from './sessionActions';
+import * as actions from '../App/sessionActions';
 
-export class LogoutFederated extends React.Component {
+export class LogoutSession extends React.Component {
   componentWillMount() {
     const { logout } = this.props;
-    logout({ federated: true });
+    logout({ federated: false });
   }
 
   render() {
@@ -21,13 +22,12 @@ export class LogoutFederated extends React.Component {
   }
 }
 
-LogoutFederated.propTypes = {
+LogoutSession.propTypes = {
   logout: PropTypes.func.isRequired,
-  authenticated: PropTypes.func.isRequired,
 };
 
 const mapDispatchToProps = {
   logout: actions.logout,
 };
 
-export default connect(undefined, mapDispatchToProps)(LogoutFederated);
+export default connect(undefined, mapDispatchToProps)(LogoutSession);
